nodefs: make everything compile on darwin
