+++ nodefs/simple_test.go
@@ -17,11 +17,8 @@ import (
 	"testing"
 	"time"
 
-	"golang.org/x/sys/unix"
-
 	"github.com/hanwen/go-fuse/fuse"
 	"github.com/hanwen/go-fuse/internal/testutil"
-	"github.com/kylelemons/godebug/pretty"
 )
 
 type testCase struct {
@@ -118,7 +115,7 @@ func TestBasic(t *testing.T) {
 	}
 
 	stat := fuse.ToStatT(fi)
-	if got, want := stat.Mode, uint32(fuse.S_IFREG|0644); got != want {
+	if got, want := uint32(stat.Mode), uint32(fuse.S_IFREG|0644); got != want {
 		t.Errorf("got mode %o, want %o", got, want)
 	}
 
@@ -164,7 +161,7 @@ func TestFile(t *testing.T) {
 	}
 
 	stat := fuse.ToStatT(fi)
-	if got, want := stat.Mode, uint32(fuse.S_IFREG|0755); got != want {
+	if got, want := uint32(stat.Mode), uint32(fuse.S_IFREG|0755); got != want {
 		t.Errorf("Fstat: got mode %o, want %o", got, want)
 	}
 
@@ -299,91 +296,6 @@ func TestRenameDestExist(t *testing.T) {
 
 func TestRenameDestNoExist(t *testing.T) {
 	testRenameOverwrite(t, false)
-}
-
-func TestRenameNoOverwrite(t *testing.T) {
-	tc := newTestCase(t, true, true)
-	defer tc.Clean()
-
-	if err := os.Mkdir(tc.origDir+"/dir", 0755); err != nil {
-		t.Fatalf("Mkdir: %v", err)
-	}
-	tc.writeOrig("file", "hello", 0644)
-	tc.writeOrig("dir/file", "x", 0644)
-
-	f1, err := syscall.Open(tc.mntDir+"/", syscall.O_DIRECTORY, 0)
-	if err != nil {
-		t.Fatalf("open 1: %v", err)
-	}
-	defer syscall.Close(f1)
-	f2, err := syscall.Open(tc.mntDir+"/dir", syscall.O_DIRECTORY, 0)
-	if err != nil {
-		t.Fatalf("open 2: %v", err)
-	}
-	defer syscall.Close(f2)
-
-	if err := unix.Renameat2(f1, "file", f2, "file", unix.RENAME_NOREPLACE); err == nil {
-		t.Errorf("rename NOREPLACE succeeded")
-	} else if err != syscall.EEXIST {
-		t.Errorf("got %v (%T) want EEXIST", err, err)
-	}
-}
-
-func TestRenameExchange(t *testing.T) {
-	tc := newTestCase(t, true, true)
-	defer tc.Clean()
-
-	if err := os.Mkdir(tc.origDir+"/dir", 0755); err != nil {
-		t.Fatalf("Mkdir: %v", err)
-	}
-	tc.writeOrig("file", "hello", 0644)
-	tc.writeOrig("dir/file", "x", 0644)
-
-	f1, err := syscall.Open(tc.mntDir+"/", syscall.O_DIRECTORY, 0)
-	if err != nil {
-		t.Fatalf("open 1: %v", err)
-	}
-	defer syscall.Close(f1)
-	f2, err := syscall.Open(tc.mntDir+"/dir", syscall.O_DIRECTORY, 0)
-	if err != nil {
-		t.Fatalf("open 2: %v", err)
-	}
-	defer syscall.Close(f2)
-
-	var before1, before2 unix.Stat_t
-	if err := unix.Fstatat(f1, "file", &before1, 0); err != nil {
-		t.Fatalf("Fstatat: %v", err)
-	}
-	if err := unix.Fstatat(f2, "file", &before2, 0); err != nil {
-		t.Fatalf("Fstatat: %v", err)
-	}
-
-	if err := unix.Renameat2(f1, "file", f2, "file", unix.RENAME_EXCHANGE); err != nil {
-		t.Errorf("rename EXCHANGE: %v", err)
-	}
-
-	var after1, after2 unix.Stat_t
-	if err := unix.Fstatat(f1, "file", &after1, 0); err != nil {
-		t.Fatalf("Fstatat: %v", err)
-	}
-	if err := unix.Fstatat(f2, "file", &after2, 0); err != nil {
-		t.Fatalf("Fstatat: %v", err)
-	}
-	clearCtime := func(s *unix.Stat_t) {
-		s.Ctim.Sec = 0
-		s.Ctim.Nsec = 0
-	}
-
-	clearCtime(&after1)
-	clearCtime(&after2)
-	clearCtime(&before2)
-	clearCtime(&before1)
-	if diff := pretty.Compare(after1, before2); diff != "" {
-		t.Errorf("after1, before2: %s", diff)
-	}
-	if !reflect.DeepEqual(after2, before1) {
-		t.Errorf("after2, before1: %#v, %#v", after2, before1)
-	}
 }
 
 func TestNlinkZero(t *testing.T) {
@@ -598,41 +510,6 @@ func TestStatFs(t *testing.T) {
 	}
 }
 
-func TestXAttr(t *testing.T) {
-	tc := newTestCase(t, true, true)
-	defer tc.Clean()
-
-	tc.writeOrig("file", "", 0644)
-
-	buf := make([]byte, 1024)
-	attr := "user.xattrtest"
-	if _, err := syscall.Getxattr(tc.mntDir+"/file", attr, buf); err == syscall.ENOTSUP {
-		t.Skip("$TMP does not support xattrs. Rerun this test with a $TMPDIR override")
-	}
-
-	if _, err := syscall.Getxattr(tc.mntDir+"/file", attr, buf); err != syscall.ENODATA {
-		t.Fatalf("got %v want ENOATTR", err)
-	}
-	value := []byte("value")
-	if err := syscall.Setxattr(tc.mntDir+"/file", attr, value, 0); err != nil {
-		t.Fatalf("Setxattr: %v", err)
-	}
-	sz, err := syscall.Getxattr(tc.mntDir+"/file", attr, buf)
-	if err != nil {
-		t.Fatalf("Getxattr: %v", err)
-	}
-	if bytes.Compare(buf[:sz], value) != 0 {
-		t.Fatalf("Getxattr got %q want %q", buf[:sz], value)
-	}
-	if err := syscall.Removexattr(tc.mntDir+"/file", attr); err != nil {
-		t.Fatalf("Removexattr: %v", err)
-	}
-
-	if _, err := syscall.Getxattr(tc.mntDir+"/file", attr, buf); err != syscall.ENODATA {
-		t.Fatalf("got %v want ENOATTR", err)
-	}
-}
-
 func TestGetAttrParallel(t *testing.T) {
 	// We grab a file-handle to provide to the API so rename+fstat
 	// can be handled correctly. Here, test that closing and
@@ -679,45 +556,3 @@ func TestGetAttrParallel(t *testing.T) {
 }
 
 // XXX test mknod.
-
-func TestCopyFileRange(t *testing.T) {
-	tc := newTestCase(t, true, true)
-	defer tc.Clean()
-
-	tc.writeOrig("src", "01234567890123456789", 0644)
-	tc.writeOrig("dst", "abcdefghijabcdefghij", 0644)
-
-	f1, err := syscall.Open(tc.mntDir+"/src", syscall.O_RDONLY, 0)
-	if err != nil {
-		t.Fatalf("Open src: %v", err)
-	}
-	defer syscall.Close(f1)
-	f2, err := syscall.Open(tc.mntDir+"/dst", syscall.O_RDWR, 0)
-	if err != nil {
-		t.Fatalf("Open dst: %v", err)
-	}
-	defer syscall.Close(f2)
-
-	srcOff := int64(5)
-	dstOff := int64(7)
-	if sz, err := unix.CopyFileRange(f1, &srcOff, f2, &dstOff, 3, 0); err != nil || sz != 3 {
-		t.Fatalf("CopyFileRange: %d,%v", sz, err)
-	}
-	if err := syscall.Close(f1); err != nil {
-		t.Fatalf("Close src: %v", err)
-	}
-	if err := syscall.Close(f2); err != nil {
-		t.Fatalf("Close dst: %v", err)
-	}
-	c, err := ioutil.ReadFile(tc.mntDir + "/dst")
-	if err != nil {
-		t.Fatalf("ReadFile: %v", err)
-	}
-
-	want := "abcdefg567abcdefghij"
-	got := string(c)
-	if got != want {
-		t.Errorf("got %q want %q", got, want)
-	}
-
-}
